remove BlogController

+++ routes/web.php
@@ -7,7 +7,7 @@ use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\admin\UesrsController;
 use App\Http\Controllers\AuthorController;
 use App\Http\Controllers\BlogController ;
-use App\Http\Controllers\Auth\PostController as ControllersPostController;
+use App\Http\Controllers\PostController as ControllersPostController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -22,13 +22,21 @@ use Illuminate\Support\Facades\Route;
 */
 
 
-Route::get('/', [BlogController::class, 'index'])->name('blog.index');
-Route::post('/fetch-data', [BlogController::class, 'fetchData'])->name('pagination.fetch_data');
-Route::post('/categories/{category_id}/fetch-data', [BlogController::class, 'fetchCategoriesData']);
-Route::post('/authors/{author_id}/fetch-data', [BlogController::class, 'fetchAuthorsData']);
-Route::get('/posts/{id}', [BlogController::class, 'post'])->name('blog.post');
-Route::get('/categories/{id}', [BlogController::class, 'category'])->name('blog.category');
-Route::get('/authors/{id}', [BlogController::class, 'author'])->name('blog.author');
+Route::get('/', [ControllersPostController::class, 'index'])->name('posts.index');
+Route::get('posts/create', [ControllersPostController::class, 'create'])->name('posts.create');
+Route::post('posts/store', [ControllersPostController::class, 'store'])->name('posts.store');
+Route::get('posts/{id}', [ControllersPostController::class, 'show'])->name('posts.show');
+Route::get('posts/{id}/edit', [ControllersPostController::class, 'edit'])->name('posts.edit');
+Route::patch('posts/{id}/update', [ControllersPostController::class, 'update'])->name('posts.update');
+Route::delete('posts/{id}/delete', [ControllersPostController::class, 'destroy'])->name('posts.destroy');
+
+
+Route::post('/fetch-data', [ControllersPostController::class, 'fetchData'])->name('pagination.fetch_data');
+Route::post('/categories/{category_id}/fetch-data', [ControllersPostController::class, 'fetchCategoriesData']);
+Route::post('/authors/{author_id}/fetch-data', [ControllersPostController::class, 'fetchAuthorsData']);
+
+Route::get('/categories/{id}', [ControllersPostController::class, 'category'])->name('posts.category');
+Route::get('/authors/{id}', [ControllersPostController::class, 'author'])->name('posts.author');
 
 
 
@@ -43,17 +51,25 @@ Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-    Route::resource('/posts', ControllersPostController::class);
-    Route::get('/my-posts', [ControllersPostController::class, 'myPosts'])->name('blog.myPosts');
+
+    Route::get('/my-posts', [ControllersPostController::class, 'myPosts'])->name('posts.myPosts');
     // route prfix admin
     Route::prefix('admin')->group(function () {
         Route::any('/users', [UesrsController::class, 'index'])->name('users.index');
         Route::resource('categories', CategoryController::class);
-        Route::resource('posts', PostController::class);
-        Route::get('posts/filterByCategory/{id}', [PostController::class, 'filterByCategory'])->name('posts.filterByCategory');
-        Route::get('posts/filterByAuthor/{id}', [PostController::class, 'filterByAuthor'])->name('posts.filterByAuthor');
+        Route::resource('posts', PostController::class)->names([
+            'index' => 'admin.posts.index',
+            'create' => 'admin.posts.create',
+            'store' => 'admin.posts.store',
+            'show' => 'admin.posts.show',
+            'edit' => 'admin.posts.edit',
+            'update' => 'admin.posts.update',
+            'destroy' => 'admin.posts.destroy',
+        ]);;
+        Route::get('posts/filterByCategory/{id}', [PostController::class, 'filterByCategory'])->name('admin.posts.filterByCategory');
+        Route::get('posts/filterByAuthor/{id}', [PostController::class, 'filterByAuthor'])->name('admin.posts.filterByAuthor');
         // Route delete post
-        Route::delete('posts/deletePost/{id}', [PostController::class, 'deletePost'])->name('posts.deletePost');
+        Route::delete('posts/deletePost/{id}', [PostController::class, 'deletePost'])->name('admin.posts.deletePost');
     });
